backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
openai==1.54.0
anthropic==0.39.0
httpx==0.27.2
celery==5.3.4
redis==5.0.1
docker==7.1.0
requests==2.32.3
urllib3==1.26.20
pyyaml==6.0.1
jinja2==3.1.2
Pillow==10.1.0
flask==3.0.0
stripe==7.0.0
markdown==3.5.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
